+++ src/assets/material.rs
@@ -1,0 +1,6 @@
+pub struct Material {
+    pub bind_group: wgpu::BindGroup,
+}
+
+#[derive(Copy, Clone, Eq, PartialEq)]
+pub struct MaterialHandle(pub u32);
