+++ src/assets/mod.rs
@@ -1,0 +1,9 @@
+pub mod loader;
+pub mod material;
+pub mod mesh;
+pub mod texture;
+
+pub use loader::load_model;
+pub use material::{Material, MaterialHandle};
+pub use mesh::{Mesh, MeshHandle, Primitive};
+pub use texture::Texture;
