+++ src/render/mod.rs
@@ -1,0 +1,9 @@
+pub mod depth;
+pub mod instance;
+pub mod uniforms;
+pub mod vertex;
+
+pub use depth::create_depth_texture;
+pub use instance::{INSTANCE_BUFFER_INITIAL_CAPACITY, InstanceData};
+pub use uniforms::{CameraUniform, LightUniform};
+pub use vertex::Vertex;
