+++ src/resources/mod.rs
@@ -1,0 +1,8 @@
+pub mod camera;
+pub mod input;
+pub mod pools;
+pub mod time;
+
+pub use camera::Camera;
+pub use input::Input;
+pub use pools::{MaterialPool, MeshPool, TexturePool};
