+++ src/systems/camera.rs
@@ -1,0 +1,1 @@
+fn camera_system() {}
